Added {co'au'a} to camxes-exp.
{co'au'a} Experimental cmavo of selma'o ZAhO.
interval event contour: at the restarting/renewal point of...
The event is being considered somewhat/approximately cyclic in nature on the large scale; the point in question is an end point for one version of going through the cycle once and the point at which the next version begins. The previous cycle need not have reached or exceeded its natural completion/conclusion. This word is useful for when a player restarts a level in a game - the choices made within the level may not be exactly the same, but much of it nonetheless is extremely similar between attempts. "Redouble" may have this meaning. Use this word when one is not at the true (last) end of an event, nor its true (first) beginning, yet is at an end and a begin.