Added {xei'e} to camxes-exp.
△ xei’e [KOhA] pro-sumti: the virtual ”here”
we are communicating in — It may feel imprecisely
metaphorical to use {vi} or {ti} or {stuzi} when talking
online and wanting to refer to ”here” (= this call,
this server, this chat room). Participants share a virtual
space, but are far apart spatially. {xei’e} captures
this spatial metaphor precisely: to {zvati} {xei’e} means
exactly ”to have access to the current communication
channel”. This metaphor then extends itself to {cliva},
{klama}, {stuzi}, {zvakansa}, saying ”{cladu} {fe} {xei’e}”,
etc.